butia_vision_bridge: removing message filters from VisionBridge
