+++ butia_recognition_pytorch/config/object_recognition_larc2021_real.yaml
@@ -1,0 +1,23 @@
+threshold: 0.3
+classes_by_category:
+    simple: [Apple, Banana, Carrot, Mayonnaise, Ketchup, Soap, Napkin, Toothpaste, Alcohol Bottle, Yeast, Chocolate Powder, Coconut Milk, Soda can, Beer can, Popcorn, Gelatine, Cracker, Chocolate Milk,]
+all_classes:
+    - Apple
+    - Banana
+    - Carrot
+    - Mayonnaise
+    - Ketchup
+    - Soap
+    - Napkin
+    - Toothpaste
+    - Alcohol Bottle
+    - Yeast
+    - Chocolate Powder
+    - Coconut Milk
+    - Soda can
+    - Beer can
+    - Popcorn
+    - Gelatine
+    - Cracker
+    - Chocolate Milk
+model_file: larc2021_real.pt
