+++ butia_recognition_pytorch/config/yolov5_network_config/weights/.gitignore
@@ -1,0 +1,4 @@
+*.pt
+*.trt
+*.onnx
+*.torchscript
