+++ butia_recognition_pytorch/launch/yolov5-simulator-foods.launch
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<launch>
+  <!-- Console launch prefix -->
+  <arg name="launch_prefix" default=""/>
+
+  <!-- Config and weights folder. -->
+  <!--arg name="yolo_weights_path"          default="$(find darknet_ros)/yolo_network_config/weights"/>
+  <arg name="yolo_config_path"           default="$(find darknet_ros)/yolo_network_config/cfg"/>
+-->
+  <!-- Load parameters -->
+  <rosparam command="load" ns="object_recognition" file="$(find object_recognition)/config/ros.yaml"/>
+  <rosparam command="load" ns="object_recognition" file="$(find object_recognition)/config/object_recognition-sim.yaml"/>
+<!--  <rosparam command="load" ns="darknet_ros" file="$(find darknet_ros)/config/ros.yaml"/>
+  <rosparam command="load" ns="darknet_ros" file="$(find darknet_ros)/config/yolov4-simulator-foods-larc2021.yaml"/>
+-->
+  <!-- Start darknet and ros wrapper -->
+  <!--node pkg="darknet_ros" type="darknet_ros" name="darknet_ros" output="screen" launch-prefix="$(arg launch_prefix)">
+    <param name="weights_path"          value="$(arg yolo_weights_path)" />
+    <param name="config_path"           value="$(arg yolo_config_path)" />
+  </node--> 
+
+  <rosparam command="load" ns="object_recognition" file="$(find object_recognition)/config/ros.yaml"/>
+  
+  <node pkg="object_recognition" type="object_recognition_node.py" name="object_recognition" output="screen"/>
+
+</launch>
