+++ butia_recognition_pytorch/scripts/yolo_recognition/__init__.py
@@ -1,0 +1,1 @@
+from .yolo_recognition import YoloRecognition
