--- object_recognition/config/yolov5_network_config/weights/.gitignore
@@ -1,4 +1,0 @@
-*.pt
-*.trt
-*.onnx
-*.torchscript
