--- object_recognition/launch/object_recognition.launch
@@ -1,14 +1,0 @@
-<launch>
-  <arg name="machine" default="localhost"/>
-  <arg name="use_machine" default="true"/>
-  <arg name="output" default="screen"/>
-  <arg name="config_file" default="object_recognition.yaml"/>
-
-  <machine name="localhost" address="localhost" if="$(arg use_machine)"/>
-
-  <rosparam command="load" ns="object_recognition" file="$(find object_recognition)/config/ros.yaml"/>
-  <rosparam command="load" ns="object_recognition" file="$(find object_recognition)/config/$(arg config_file)"/>
-  
-  <node pkg="object_recognition" type="object_recognition_node.py" output="$(arg output)" name="object_recognition" machine="$(arg machine)"/>
-
-</launch>
