--- object_recognition/scripts/yolo_recognition/__init__.py
@@ -1,1 +1,0 @@
-from .yolo_recognition import YoloRecognition
